+++ wp-content/plugins/wp-gdpr-compliance/uninstall.php
@@ -29,6 +29,7 @@ $wpdb->query("DROP TABLE IF EXISTS `{$wpdb->base_prefix}wpgdprc_consents`");
 
 // Cronjobs
 wp_clear_scheduled_hook('wpgdprc_deactivate_access_requests');
+wp_clear_scheduled_hook('wpgdprc_anonymise_requests');
 
 // Clear any cached data that has been removed
 wp_cache_flush();